Proportional Gain first try

+++ Roboter_GUI/mainwindow.cpp
@@ -34,7 +34,7 @@ void MainWindow::setGuiConnects()
     connect(ui->btn_stop, SIGNAL(clicked()), m_pMobilePlatform, SLOT(slot_stopMotion()));
     connect(ui->btn_stop, SIGNAL(clicked()), m_pMobilePlatform, SLOT(slot_endFollowLine()));
 
-    connect(ui->btn_speed, SIGNAL(valueChanged(int)), m_pMobilePlatform, SLOT(slot_setSpeed(int)));
+    connect(ui->btn_speed, SIGNAL(valueChanged(double)), m_pMobilePlatform, SLOT(slot_setSpeed(double)));
 
     connect(ui->btn_followLine, SIGNAL(clicked()), m_pMobilePlatform, SLOT(slot_followLine()));
 }
